chore(file): refactor into separate files
